get rid of (most) compiler warnings
git-svn-id: https://urjtag.svn.sourceforge.net/svnroot/urjtag/trunk@733 b68d4a1b-bc3d-0410-92ed-d4ac073336b7

+++ jtag/src/readmem.c
@@ -44,7 +44,7 @@ void
 readmem( bus_t *bus, FILE *f, uint32_t addr, uint32_t len )
 {
 	uint32_t step;
-	uint64_t a;
+	uint32_t a;
 	int bc = 0;
 #define BSIZE 4096
 	uint8_t b[BSIZE];
